requirements.txt
GitPython==3.1.44
fastmcp==2.4.0